tests/requirements.txt
pytest>=7.0
python-rtmidi>=1.5
mido>=1.3
pyserial>=3.5
